+++ Person Specific Stuff/Alex/Networking tests/Networking Experiments/src/rts/elements/Unit.java
@@ -1,16 +1,16 @@
 package rts.elements;
 
 public class Unit extends ControllableObject {
-	private double radius;
-	private double attackRange;
+	private float radius;
+	private float attackRange;
 	private int damage;
-	private double speed;
+	private float speed;
 	
 	private boolean moving;
 	private boolean attacking;
 	private boolean idling;
 	
-	public Unit(int id, double x, double y, int speed, int hp, double attackRange, int damage) {
+	public Unit(int id, float x, float y, int speed, int hp, float attackRange, int damage) {
 		this.id = id;
 		this.x = x;
 		this.y = y;
@@ -21,7 +21,7 @@ public class Unit extends ControllableObject {
 		idling = true;
 	}
 	
-	public void moveTo(int x, int y) {
+	public void moveTo(float x, float y) {
 		
 	}
