+++ Person Specific Stuff/Alex/Networking tests/Networking Experiments/src/rts/elements/structures/CommandCenter.java
@@ -3,7 +3,7 @@ package rts.elements.structures;
 import rts.elements.Building;
 
 public class CommandCenter extends Building {
-	public CommandCenter(int id, double x, double y) {
+	public CommandCenter(int id, float x, float y) {
 		super(id, x, y, 5, 5, 1000);
 	}
 }
