+++ Person Specific Stuff/Alex/Networking tests/Networking Experiments/src/rts/elements/structures/Factory.java
@@ -3,7 +3,7 @@ package rts.elements.structures;
 import rts.elements.Building;
 
 public class Factory extends Building {
-	public Factory(int id, double x, double y) {
+	public Factory(int id, float x, float y) {
 		super(id, x, y, 4, 3, 300);
 	}
 }
