+++ Person Specific Stuff/Alex/Networking tests/Networking Experiments/src/rts/elements/structures/Generator.java
@@ -3,7 +3,7 @@ package rts.elements.structures;
 import rts.elements.Building;
 
 public class Generator extends Building {
-	public Generator(int id, double x, double y) {
+	public Generator(int id, float x, float y) {
 		super(id, x, y, 3, 4, 400);
 	}
 }
