+++ Person Specific Stuff/Alex/Networking tests/Networking Experiments/src/rts/elements/units/Miner.java
@@ -5,10 +5,10 @@ import rts.elements.Unit;
 public class Miner extends Unit {
 	private static final int SPEED = 2;
 	private static final int HP = 50;
-	private static final double RANGE = 0.2;
+	private static final float RANGE = 0.2f;
 	private static final int DAMAGE = 5;
 	
-	public Miner(int id, double x, double y) {
+	public Miner(int id, float x, float y) {
 		super(id, x, y, SPEED, HP, RANGE, DAMAGE);
 	}
 }
